+++ react/FormInput-dropdown.tsx
@@ -4,5 +4,5 @@ import { DropdownInput } from './components/DropdownRenderer'
 import { BaseInputProps } from './typings/InputProps'
 
 export default function FormInput(props: BaseInputProps) {
-  return <DropdownInput path={props.path} />
+  return <DropdownInput pointer={props.pointer} />
 }