+++ react/__mocks__/vtex.css-handles.js
@@ -1,0 +1,8 @@
+import React from 'react'
+
+export const useCssHandles = handles => {
+  return handles.reduce((acc, key) => {
+    acc[key] = key
+    return acc
+  }, {})
+}