+++ react/components/CheckboxRenderer.tsx
@@ -2,8 +2,8 @@ import React, { FC } from 'react'
 import { Checkbox } from 'vtex.styleguide'
 import { UseCheckboxReturnType, useCheckbox } from 'react-hook-form-jsonschema'
 
-export const CheckboxInput: FC<{ path: string }> = props => {
-  const checkboxObject = useCheckbox(props.path)
+export const CheckboxInput: FC<{ pointer: string }> = props => {
+  const checkboxObject = useCheckbox(props.pointer)
   return <CheckboxRenderer checkboxObject={checkboxObject} />
 }
 
@@ -13,7 +13,7 @@ export const CheckboxRenderer: FC<{
   const checkboxObject = props.checkboxObject
 
   if (checkboxObject.isSingle) {
-    const checked = checkboxObject.formContext.watch(checkboxObject.path)
+    const checked = checkboxObject.formContext.watch(checkboxObject.pointer)
 
     return (
       <>
@@ -22,13 +22,11 @@ export const CheckboxRenderer: FC<{
           label={checkboxObject.getObject().title}
           required={checkboxObject.isRequired}
           value="true"
-          {...(checkboxObject.formContext.watch(checkboxObject.path)
-            ? { checked: true }
-            : { checked: false })}
+          {...(checked ? { checked: true } : { checked: false })}
           onChange={() => {
-            const path = checkboxObject.path
+            const pointer = checkboxObject.pointer
 
-            checkboxObject.formContext.setValue(path, !checked)
+            checkboxObject.formContext.setValue(pointer, !checked)
           }}
           checked={Boolean(checked)}
         />
@@ -40,7 +38,7 @@ export const CheckboxRenderer: FC<{
     <>
       {checkboxObject.getItems().map((value, index) => {
         const checked = checkboxObject.formContext.watch(
-          `${checkboxObject.path}[${index}]`
+          `${checkboxObject.pointer}[${index}]`
         )
         return (
           <Checkbox
@@ -49,13 +47,11 @@ export const CheckboxRenderer: FC<{
             label={checkboxObject.getObject().title}
             required={checkboxObject.isRequired}
             value={value}
-            {...(checkboxObject.formContext.watch(checkboxObject.path)
-              ? { checked: true }
-              : { checked: false })}
+            {...(checked ? { checked: true } : { checked: false })}
             onChange={() => {
-              const path = checkboxObject.path
+              const pointer = checkboxObject.pointer
 
-              checkboxObject.formContext.setValue(path, !checked)
+              checkboxObject.formContext.setValue(pointer, !checked)
             }}
             checked={Boolean(checked)}
           />