+++ CHANGELOG.md
@@ -13,6 +13,10 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 - Error messages when submitting goes wrong.
 - Success component(`form-success`) when the form is succesfully submitted.
 
+### Changed
+
+- props of `form-input` and `form-field-group` from `path` to `pointer`. And changed how to indicate the root path and subschema paths.
+
 ## [0.1.0] - 2020-02-04
 
 ### Added